setting is read from ini file. only modified file from previous execution is to copy .

+++ file_copy/src/main.rs
@@ -44,19 +44,38 @@ fn check_arguments(args: &Vec<String>) -> Result<Arguments, String> {
     Ok(arguments)
 }
 
-fn get_command_name(path: String) -> String {
-    let mut command_name = String::new();
-    let i = Ini::load_from_file(path + "\\conf.ini").unwrap();
-    for (sec, prop) in i.iter() {
+fn get_ini(path: &String) -> Ini {
+    let ini_file = Ini::load_from_file(get_ini_path(path)).unwrap();
+    ini_file
+}
+
+fn get_ini_path(path: &String) -> String {
+    path.clone() + "\\conf.ini"
+}
+
+fn get_ini_value(path: String, key: String) -> String {
+    let mut value = String::new();
+    let ini_file = get_ini(&path);
+    for (sec, prop) in ini_file.iter() {
         if *sec == Some("kihon".to_string()) {
             for (k, v) in prop.iter() {
-                if *k == "command" {
-                     command_name = String::from_str(&*v.to_string()).expect("failed read ini file.");
+                if *k == key {
+                     value = String::from_str(&*v.to_string()).expect("failed read ini file.");
                 }
             }
         }
     }
-    command_name
+    value
+}
+
+fn set_ini_modified_date(path: String, date: DateTime<Local>) {
+    let conf_path = get_ini_path(&path);
+    let mut ini_file = get_ini(&path);
+    ini_file.set_to(Some("kihon"), "lastmoddate".to_string(), date.to_string());
+    match ini_file.write_to_file(&conf_path) {
+        Err(_) => println!("conf.ini write error"),
+        Ok(_) => println!("conf.ini write {:?}", date.to_string()),
+    }
 }
 
 fn main() {
@@ -81,9 +100,21 @@ fn main() {
     };
 
     //1日前をセット
-    let mut now: DateTime<Local> = std::time::SystemTime::now().into();
-    now = now - chrono::Duration::days(1);
-    println!("{:?}", now);
+    let date = get_ini_value(exe_path.clone(), "lastmoddate".to_string());
+    let last_update_date: DateTime<Local> = match DateTime::parse_from_str(&date, "%Y-%m-%d %H:%M:%S%.6f %z") {
+        Ok(date) => {
+            Local.datetime_from_str(&date.to_string(), "%Y-%m-%d %H:%M:%S%.6f %z").unwrap()
+        },
+        Err(e) => {
+            println!("datetime parse error {:?}", e);
+            let mut now = std::time::SystemTime::now().into();
+            now = now - chrono::Duration::days(1);
+            now
+        },
+    };
+
+    println!("target date is over {:?}", last_update_date);
+    let mut most_newest_date = last_update_date.clone();
 
     //取込対象ファイルをコピー
     match read_dir(arguments.path_from) {
@@ -100,8 +131,13 @@ fn main() {
                         Err(e) => panic!("file property not read. {:?}", e)
                     };
                     
-                    //更新日が１日前以降が処理対象
-                    if now < mod_date {
+                    //前回に処理した最終更新日より新しいファイルを対象とする
+                    if last_update_date < mod_date {
+                        //より新しい日付を保存する。
+                        if most_newest_date < mod_date {
+                            most_newest_date = mod_date;
+                        }
+
                         println!("copy target {:?}", &p);
                         arguments.path_to.push(&p.file_name().unwrap());
                         println!("copy dest {:?}", &arguments.path_to.as_path());
@@ -118,24 +154,29 @@ fn main() {
         },
     }
     
-    let command_name = get_command_name(exe_path);
+
+    //最も新しかった日付を保存
+    set_ini_modified_date(exe_path.clone(), most_newest_date);
+
     //外部プログラム呼び出し
-    Command::new(command_name)
-        .arg(&arguments.path_to)
-        .output().
-        expect("failed to execute process");
+    let command_name = get_ini_value(exe_path.clone(), "command".to_string());
+    if !command_name.is_empty() {
+        Command::new(command_name)
+            .arg(&arguments.path_to)
+            .output().
+            expect("failed to execute process");
 
-
-    //コースを作成したのでコピーしたファイルを削除
-    match read_dir(arguments.path_to) {
-        Err(e) => println!("{:?}", e.kind()),
-        Ok(paths) => {
-            for path in paths {
-                let p = path.unwrap().path();
-                if p.is_file() {
-                    match remove_file(&p) {
-                        Ok(()) => println!("deleted file at {:?}", &p),
-                        Err(_) => println!("fail delete file {:?}", &p),
+        //コピーしたファイルを削除
+        match read_dir(arguments.path_to) {
+            Err(e) => println!("{:?}", e.kind()),
+            Ok(paths) => {
+                for path in paths {
+                    let p = path.unwrap().path();
+                    if p.is_file() {
+                        match remove_file(&p) {
+                            Ok(()) => println!("deleted file at {:?}", &p),
+                            Err(_) => println!("fail delete file {:?}", &p),
+                        }
                     }
                 }
             }
